+++ elasticsearch_dsl/search.py
@@ -8,7 +8,7 @@ from elasticsearch.exceptions import TransportError
 
 from .query import Q, EMPTY_QUERY, Bool
 from .aggs import A, AggBase
-from .utils import DslBase
+from .utils import DslBase, AttrDict
 from .response import Response, Hit, SuggestResponse
 from .connections import connections
 
@@ -42,6 +42,12 @@ class QueryProxy(object):
             self._proxied = Q(self._proxied.to_dict())
             setattr(self._proxied, attr_name, value)
         super(QueryProxy, self).__setattr__(attr_name, value)
+
+    def __getstate__(self):
+        return (self._search, self._proxied, self._attr_name)
+
+    def __setstate__(self, state):
+        self._search, self._proxied, self._attr_name = state
 
 
 class ProxyDescriptor(object):
@@ -95,6 +101,15 @@ class Request(object):
 
         self._params = {}
         self._extra = extra or {}
+
+    def __eq__(self, other):
+        return (
+            isinstance(other, Request) and
+            other._params == self._params and
+            other._index == self._index and
+            other._doc_type == self._doc_type and
+            other.to_dict() == self.to_dict()
+        )
 
     def params(self, **kwargs):
         """
@@ -226,6 +241,9 @@ class Search(Request):
 
     def filter(self, *args, **kwargs):
         return self.query(Bool(filter=[Q(*args, **kwargs)]))
+
+    def exclude(self, *args, **kwargs):
+        return self.query(Bool(filter=[~Q(*args, **kwargs)]))
 
     def __iter__(self):
         """
@@ -471,7 +489,7 @@ class Search(Request):
     def highlight(self, *fields, **kwargs):
         """
         Request highlighting of some fields. All keyword arguments passed in will be
-        used as parameters. Example::
+        used as parameters for all the fields in the ``fields`` parameter. Example::
 
             Search().highlight('title', 'body', fragment_size=50)
 
@@ -481,6 +499,20 @@ class Search(Request):
                 "highlight": {
                     "fields": {
                         "body": {"fragment_size": 50},
+                        "title": {"fragment_size": 50}
+                    }
+                }
+            }
+
+        If you want to have different options for different fields you can call ``highlight`` twice::
+
+            Search().highlight('title', fragment_size=50).highlight('body', fragment_size=100)
+
+        which will produce::
+            {
+                "highlight": {
+                    "fields": {
+                        "body": {"fragment_size": 100},
                         "title": {"fragment_size": 50}
                     }
                 }
@@ -623,10 +655,33 @@ class Search(Request):
                 doc_type=self._doc_type,
                 **self._params
             ):
-            yield self._doc_type_map.get(hit['_type'], Hit)(hit)
+            callback = self._doc_type_map.get(hit['_type'], Hit)
+            callback = getattr(callback, 'from_es', callback)
+            yield callback(hit)
+
+    def delete(self):
+        """
+        delete() executes the query by delegating to delete_by_query()
+        """
+
+        es = connections.get_connection(self._using)
+
+        return AttrDict(
+            es.delete_by_query(
+                index=self._index,
+                body=self.to_dict(),
+                doc_type=self._doc_type,
+                **self._params
+            )
+        )
+
 
 
 class MultiSearch(Request):
+    """
+    Combine multiple :class:`~elasticsearch_dsl.Search` objects into a single
+    request.
+    """
     def __init__(self, **kwargs):
         super(MultiSearch, self).__init__(**kwargs)
         self._searches = []
@@ -643,6 +698,13 @@ class MultiSearch(Request):
         return ms
 
     def add(self, search):
+        """
+        Adds a new :class:`~elasticsearch_dsl.Search` object to the request::
+
+            ms = MultiSearch(index='my-index')
+            ms = ms.add(Search(doc_type=Category).filter('term', category='python'))
+            ms = ms.add(Search(doc_type=Blog))
+        """
         ms = self._clone()
         ms._searches.append(search)
         return ms
@@ -663,6 +725,9 @@ class MultiSearch(Request):
         return out
 
     def execute(self, ignore_cache=False, raise_on_error=True):
+        """
+        Execute the multi search request and return a list of search results.
+        """
         if ignore_cache or not hasattr(self, '_response'):
             es = connections.get_connection(self._using)